// microservices/user-service/src/types/enums.ts
export enum RoleUtilisateur {
  ADMIN = "admin",
  ANETI = "aneti",
  INSPECTEUR = "inspecteur",
  CANDIDAT = "candidat",
  ENTREPRISE = "entreprise",
}

export enum StatutUtilisateur {
  EN_ATTENTE = "en_attente",
  APPROUVE = "approuve",
  ACTIF = "actif",
  REFUSE = "refuse",
  SUSPENDU = "suspendu",
  INACTIF = "inactif",
}

export enum GenreUtilisateur {
  HOMME = "homme",
  FEMME = "femme",
}

export enum StatutValidationEntreprise {
  VALIDE = "valide",
  INVALIDE = "invalide",
}